# Conflicts:
#	datacaptureapp/views.py

+++ datacaptureapp/views.py
@@ -1,3 +1,10 @@
+from django.core.exceptions import PermissionDenied
+from django.http import QueryDict, HttpResponse, JsonResponse
+from django.core import serializers
+from django.http import FileResponse, QueryDict, HttpResponse, JsonResponse
+from django.shortcuts import render, redirect, get_object_or_404
+from django.urls import reverse
+
 from django.http import QueryDict, HttpResponse, JsonResponse, HttpResponseServerError
 from django.shortcuts import render, redirect
 from datacaptureapp.forms import *
@@ -50,17 +57,42 @@ def project(request, pk=0):
     :param pk: The id of the project
     :return:
     """
-    if pk == 0:
-        return HttpResponseServerError()
-    requested_project = Project.objects.get(id=pk)
-    geojson = generate_geojson(requested_project)
-    owner = requested_project.user.all().first()
-    attributes = Attribute.objects.filter(project__id=pk)
-    data = Data.objects.filter(attribute__in=attributes)
-    requested_nodes = Node.objects.filter(project_id=pk)
-    overview = get_node_overview(data, requested_nodes)
-    return render(request, 'datacaptureapp/Project.html',
-                  {'project': requested_project, 'owner': owner, 'geojson': geojson, 'overview': overview})
+    requested_project = Project.objects.filter(id=pk).first()
+    if not requested_project:
+        if request.headers.get('search-by-id'):
+            messages.error(request, 'The project you were looking for does not exist')
+            return JsonResponse({"messages": messagesToList(request)})
+        else:
+            raise PermissionDenied
+    else:
+        if requested_project.is_public or requested_project.user.filter(email=request.user.email).first():
+            # POST request to change private/public
+            if request.POST:
+                form = ChangePublicPrivateForm(request.POST, instance=requested_project)
+                if form.is_valid():
+                    form.save()
+                    return JsonResponse({'is_public': requested_project.is_public})
+            else:
+                if request.headers.get('search-by-id'):
+                    return JsonResponse({'url': '/projects/' + str(pk) + '/'})
+                else:
+                    geojson = generate_geojson(requested_project)
+                    owner = requested_project.user.all().first()  # TODO there are more users now, we do not specify the owner
+                    attributes = Attribute.objects.filter(project__id=pk)
+                    data = Data.objects.filter(attribute__in=attributes)
+                    requested_nodes = Node.objects.filter(project_id=pk)
+                    overview = get_node_overview(data, requested_nodes)
+                    return render(request, 'datacaptureapp/Project.html',
+                                  {'project': requested_project, 'owner': owner, 'geojson': geojson,
+                                   'overview': overview, 'data': data, 'attributes': attributes,
+                                   'requested_nodes': requested_nodes})
+        else:
+            if request.headers.get('search-by-id'):
+                messages.error(request, 'This project is private. Ask the project owner to add you to this project.')
+                return JsonResponse({"messages": messagesToList(request)})
+            else:
+                raise PermissionDenied
+
 
 
 def get_node_overview(data, requested_nodes):
@@ -79,6 +111,7 @@ def get_node_overview(data, requested_nodes):
     return overview
 
 
+
 @login_required()
 def addnode(request, pk):
     """
@@ -87,39 +120,44 @@ def addnode(request, pk):
     :param pk: The id of the project
     :return:
     """
-    requested_project = Project.objects.filter(id=pk).first()
-    attributes = Attribute.objects.filter(project=requested_project)
-    if request.method == "POST":
-        latitude_formatted = "{:.8f}".format(Decimal(request.POST.get('latitude')))
-        longitude_formatted = "{:.8f}".format(Decimal(request.POST.get('longitude')))
-        request.POST._mutable = True
-        request.POST['latitude'] = latitude_formatted
-        request.POST['longitude'] = longitude_formatted
-        request.POST._mutable = False
-        node_form = CreateNodeForm(request.POST, request.FILES)
-        if node_form.is_valid():
-            node = node_form.save(commit=False)
-            node.project = requested_project
-            node.save()
-        for attribute in attributes:
-            data_query_dict = QueryDict('value=' + request.POST.get(attribute.name))
-            data_form = CreateDataForm(data_query_dict)
-            if not data_form.is_valid():
-                data_form = CreateDataForm(QueryDict("value=Null"))
-            data = data_form.save(commit=False)
-            data.node = node
-            data.attribute = attribute
-            data.save()
-        return redirect('project', pk)
+    requested_project = get_object_or_404(Project, pk=pk)
+    if requested_project.is_public or requested_project.user.filter(email=request.user.email).first():
+        attributes = Attribute.objects.filter(project=requested_project)
+        if request.method == "POST":
+            latitude_formatted = "{:.8f}".format(Decimal(request.POST.get('latitude')))
+            longitude_formatted = "{:.8f}".format(Decimal(request.POST.get('longitude')))
+            request.POST._mutable = True
+            request.POST['latitude'] = latitude_formatted
+            request.POST['longitude'] = longitude_formatted
+            request.POST._mutable = False
+            node_form = CreateNodeForm(request.POST, request.FILES)
 
-    node_form = CreateNodeForm()
-    datas = []
-    for attribute in attributes:
-        data = CreateDataForm(QueryDict('value=Null'))
-        data = data.save(commit=False)
-        data.attribute = attribute
-        datas.append(data)
-    return render(request, 'datacaptureapp/AddFeature.html', {'node_form': node_form, 'datas': datas, 'project_id': pk})
+            if node_form.is_valid():
+                node = node_form.save(commit=False)
+                node.project = requested_project
+                node.save()
+
+            for attribute in attributes:
+                data_query_dict = QueryDict('value=' + request.POST.get(attribute.name))
+                data_form = CreateDataForm(data_query_dict)
+                if not data_form.is_valid():
+                    data_form = CreateDataForm(QueryDict("value=Null"))
+                data = data_form.save(commit=False)
+                data.node = node
+                data.attribute = attribute
+                data.save()
+            return redirect('project', pk)
+        else:
+            node_form = CreateNodeForm()
+            datas = []
+            for attribute in attributes:
+                data = CreateDataForm(QueryDict('value=Null'))
+                data = data.save(commit=False)
+                data.attribute = attribute
+                datas.append(data)
+            return render(request, 'datacaptureapp/AddFeature.html', {'node_form': node_form, 'datas': datas, 'project_id': pk})
+    else:
+        raise PermissionDenied
 
 
 @login_required()
@@ -130,39 +168,44 @@ def nodes(request, pk):
     :param pk: The id of the project
     :return:
     """
-    if request.method == 'POST':
-        post = request.POST
-        if 'data_type' in post:
-            data_type = request.POST.get('data_type')
-            project = Project.objects.get(id=pk)
-            if data_type == 'CSV':
-                response = HttpResponse(content_type='text/csv')
-                response['Content-Disposition'] = 'attachment; filename="{}.csv"'.format(project.name)
-                generate_csv(response, project)
-            elif data_type == 'GeoJSON':
-                geojson = generate_geojson(project)
-                response = HttpResponse(content_type='application/json')
-                response['Content-Disposition'] = 'attachment; filename="{}.geojson"'.format(project.name)
-                response.write(geojson)
-            elif data_type == 'Excel':
-                response = HttpResponse(content_type='application/vnd.ms-excel')
-                response['Content-Disposition'] = "attachment; filename={}.xlsx".format(project.name)
-                output = generate_xls(project)
-                response.write(output)
-            else:
-                response = HttpResponseServerError('<h1>Something went wrong</h1>')
-            return response
-        elif 'remove_node' in post:
-            Node.objects.get(id=post['remove_node']).delete()
-    attributes = Attribute.objects.filter(project__id=pk)
-    data = Data.objects.filter(attribute__in=attributes)
-    requested_nodes = Node.objects.filter(project_id=pk)
-    overview = get_node_overview(data, requested_nodes)
-    images = {}
-    for node in requested_nodes:
-        images[node.pk] = node.picture
-    return render(request, 'datacaptureapp/FeatureOverview.html',
-                  {'overview': overview, 'images': images, 'attributes': attributes})
+    requested_project = get_object_or_404(Project, pk=pk)
+    if requested_project.is_public or requested_project.user.filter(email=request.user.email).first():
+        if request.method == 'POST':
+            post = request.POST
+            if 'data_type' in post:
+                data_type = request.POST.get('data_type')
+                if data_type == 'CSV':
+                    response = HttpResponse(content_type='text/csv')
+                    response['Content-Disposition'] = 'attachment; filename="{}.csv"'.format(requested_project.name)
+                    generate_csv(response, requested_project)
+                elif data_type == 'GeoJSON':
+                    geojson = generate_geojson(requested_project)
+                    response = HttpResponse(content_type='application/json')
+                    response['Content-Disposition'] = 'attachment; filename="{}.geojson"'.format(json.loads(geojson)['name'])
+                    response.write(geojson)
+                elif data_type == 'Excel':
+                    response = HttpResponse(content_type='application/vnd.ms-excel')
+                    response['Content-Disposition'] = "attachment; filename={}.xlsx".format(project.name)
+                    output = generate_xls(project)
+                    response.write(output)
+                else:
+                    response = HttpResponseServerError('<h1>Something went wrong</h1>')
+                return response
+            elif 'remove_node' in post:
+                Node.objects.get(id=post['remove_node']).delete()
+        else:
+            attributes = Attribute.objects.filter(project__id=pk)
+            data = Data.objects.filter(attribute__in=attributes)
+            requested_nodes = Node.objects.filter(project_id=pk)
+            overview = get_node_overview(data, requested_nodes)
+            images = {}
+            for node in requested_nodes:
+                images[node.pk] = node.picture
+            return render(request, 'datacaptureapp/FeatureOverview.html',
+                          {'overview': overview, 'images': images, 'attributes': attributes})
+    else:
+        raise PermissionDenied
+
 
 
 @login_required()
@@ -174,27 +217,32 @@ def edit_node(request, pk, nk):
     :param nk: The id of the node
     :return:
     """
-    if request.method == 'POST':
-        post = request.POST
+    requested_project = get_object_or_404(Project, pk=pk)
+    node = get_object_or_404(Node, id=nk)
+    if requested_project.is_public or requested_project.user.filter(email=request.user.email).first():
+        if request.method == 'POST':
+            post = request.POST
+            for coor in ['longitude', 'latitude']:
+                value = post[coor]
+                if value != '':
+                    node.coor = value
+            if 'picture' in request.FILES:
+                node.picture = request.FILES['picture']
+            node.save()
+            attributes = Attribute.objects.filter(project=Project.objects.get(id=pk))
+            for attribute in attributes:
+                value = post[attribute.name]
+                if value != '':
+                    data = Data.objects.get(node=node, attribute=attribute)
+                    data.value = value
+                    data.save()
+            return redirect('nodes', pk)
         node = Node.objects.get(id=nk)
-        for coor in ['longitude', 'latitude']:
-            value = post[coor]
-            if value != '':
-                node.coor = value
-        if 'picture' in request.FILES:
-            node.picture = request.FILES['picture']
-        node.save()
-        attributes = Attribute.objects.filter(project=Project.objects.get(id=pk))
-        for attribute in attributes:
-            value = post[attribute.name]
-            if value != '':
-                data = Data.objects.get(node=node, attribute=attribute)
-                data.value = value
-                data.save()
-        return redirect('nodes', pk)
-    node = Node.objects.get(id=nk)
-    datas = Data.objects.filter(node=node)
-    return render(request, 'datacaptureapp/EditNode.html', {'node': node, 'datas': datas})
+        datas = Data.objects.filter(node=node)
+        return render(request, 'datacaptureapp/EditNode.html', {'node': node, 'datas': datas})
+    else:
+        raise PermissionDenied
+
 
 
 @login_required()
@@ -205,17 +253,20 @@ def add_attribute(request, pk):
     :param pk: The id of the project
     :return:
     """
-    if request.method == 'POST':
-        form = CreateAttributeForm(request.POST)
-        if form.is_valid():
-            new_attribute = form.save(commit=False)
-            project = Project.objects.filter(id=pk).first()
-            new_attribute.project = project
-            new_attribute.save()
-            return redirect('attributes', pk)
+    requested_project = get_object_or_404(Project, pk=pk)
+    if requested_project.is_public or requested_project.user.filter(email=request.user.email).first():
+        if request.method == 'POST':
+            form = CreateAttributeForm(request.POST)
+            if form.is_valid():
+                new_attribute = form.save(commit=False)
+                new_attribute.project = requested_project
+                new_attribute.save()
+                return redirect('attributes', pk)
+        else:
+            form = CreateAttributeForm
+            return render(request, 'datacaptureapp/FormCreation.html', {'form': form})
     else:
-        form = CreateAttributeForm
-        return render(request, 'datacaptureapp/FormCreation.html', {'form': form})
+        raise PermissionDenied
 
 
 @login_required()
@@ -242,25 +293,28 @@ def team(request, pk):
     :param request: The incoming request
     :return:
     """
-    requested_project = Project.objects.filter(pk=pk).first()
-    team_members = requested_project.user.all()
-    if request.POST:
-        form = AddMemberForm(request.POST)
-        if form.is_valid():
-            if Account.objects.filter(email=request.POST.get('email')).exists():
-                account = Account.objects.filter(email=form.cleaned_data.get('email')).first()
-                requested_project.user.add(account)
-                messages.success(request, 'Successfully added the user to this project')
-                return JsonResponse({"messages": messagesToList(messages, request), 'email': account.email,
-                                     'username': account.username})
-            else:
-                messages.error(request, 'Adding team member failed: no user found with that email')
-                return JsonResponse({"messages": messagesToList(messages, request)})
+    requested_project = get_object_or_404(Project, pk=pk)
+    if requested_project.user.filter(email=request.user.email).first():
+        team_members = requested_project.user.all()
+        if request.POST:
+            form = AddMemberForm(request.POST)
+            if form.is_valid():
+                if Account.objects.filter(email=request.POST.get('email')).exists():
+                    account = Account.objects.filter(email=form.cleaned_data.get('email')).first()
+                    requested_project.user.add(account)
+                    messages.success(request, 'Successfully added the user to this project')
+                    return JsonResponse({"messages": messagesToList(request), 'email': account.email,
+                                         'username': account.username})
+                else:
+                    messages.error(request, 'Adding team member failed: no user found with that email')
+                    return JsonResponse({"messages": messagesToList(request)})
 
+        else:
+            form = AddMemberForm()
+            return render(request, 'datacaptureapp/ProjectTeam.html',
+                          {'form': form, 'team': team_members, 'project': requested_project})
     else:
-        form = AddMemberForm()
-        return render(request, 'datacaptureapp/ProjectTeam.html',
-                      {'form': form, 'team': team_members, 'project': requested_project})
+        raise PermissionDenied
 
 
 @login_required()
